Update patch set 4
Patch Set 4:

Hi Sridhar,

the changes are so small, that it would be better to fix it as part of this patch. I would say, that separate ticket is not needed and Otto could cover it with next patch version.

Hi Otto, do you agree?

Best Regards,
Martin

Patch-set: 4
Label: Verified=0