IETF Draft: Comments from IETF-96
JIRA: VSPERF-410

x Ramki on OVS features - the scope of benchmarking does not include
the ever-growing list of OVS features, only the general and
switch-agnostic features will be assessed.

x Justify the length of SOAK tests (with variability allowed),
mention that the goal is stability characterization, not
the typical short term benchmarks of performance.

x Fix the LTD Reference to point to Brahmaputra version
(which is frozen in the release docs)

Change-Id: I2f795cd42e7ec2e06e5cd35bbf0c1f336a5e43ca
Signed-off-by: Al Morton <acmorton@att.com>
Reviewed-by: Maryam Tahhan <maryam.tahhan@intel.com>
Reviewed-by: Billy O'Mahony<billy.o.mahony@intel.com>
Reviewed-by: Christian Trautman <ctrautma@redhat.com>