Update patch set 1
Patch Set 1:

Please have a look at jira VSPERF-415 for details and explanation.

I'll ask opnfvdoc team, if there is any way to use RST internal references to avoid html links. So far it seems, that with current way of docs "compilation" it is not possible to make a :ref: to another docs directory. However I might be missing something.

Patch-set: 1