Update patch set 1
Patch Set 1:

(2 comments)

Hi Martin,

A few comments below on the code. Nothing big honestly. 

Would it be cumbersome since we are now forcing test params to be in all Caps to just force the param name to uppercase in the code to remove this requirement? It might make it easier, also I sometimes find when testing that if I mistype a test-param name I get no warning. I don't see that I did this until I look over the logs or find the behavior wasn't what I expected. This could be more of a long term goal that if a test param value doesn't match anything in the conf files the system might warn you. Probably outside the scope of this patch but I just wanted to mention it since you are working in this code space right now. I don't know if we require test params to be in conf files exclusively before allowing them to be passed on the CLI.

Lastly, I will try to test this early next week and give feedback if I see any issues.

Thanks.

Patch-set: 1
Label: Code-Review=0