Update patch set 1
Patch Set 1:

(2 comments)

Hi Christian,

thanks for a review and useful hints.

I would like to keep UPPER CASE enforcement of configuration parameters. It should help any user to detect, that there has been a change of handling of configuration parameters.

I've also implemented a simple check of configuration parameter names. So in case, that parameter name is misspelled, etc. then vsperf execution will fail.

I was thinking also about a comparison of old and new data types, but it won't be that straight forward, so I've abandoned this idea for now. It means, that only toplevel type (consider complex dictionaries) would be checked and in case of int/long/float mismatch it is probably not good to raise an error, but rather show a warning. What do you think?

Patch-set: 1
Label: Verified=0