Update patch set 4
Patch Set 4:

Hi Christian,

good catch! The old algorithm for parameter parsing was not able to process multiple equal signs. I modified it a bit and I tried to keep it simple. So far it seems, that it covers all cases I tried. Let's see if it will pass your testing.

--Martin

Patch-set: 4