Update patch set 1
Patch Set 1:

Hi Martin, 

Looks like we'll have to make a few fixes in the Xena code to support this. The conversion of some values to text will make some of the algorithms fail in the Xena code. We'll have to do some typecasting. I'll send you an email with my current failure which should be easy to fix.

Patch-set: 1
Label: Code-Review=0