Update patch set 1
Patch Set 1:

Hi Antonio,

I've generated documentation with official opnfvdocs scripts and generation went smoothly without errors or warnings. Also generated document for this review looks ok and links to the Default test parameters works correctly. Is it possible, that the issue is related to the way how online editor handles end of lines?

In general, we often break long links to several lines, to keep rst source file with reasonable long lines. So either we have to change it in whole documentation or to keep it as it is (my preference).

Anyhow, while checking your comments, I've found, that one link was missing in original docs, so I'll add it in next patch version.

Patch-set: 1
Label: Verified=0