testpmd_pvp_fix: Fix SR-IOV QemuPciPassthrough mode to not use vdev
A recent patch introduced a bug where SR-IOV would use code designed
for use with testpmd_pvp scenarios. This is because this scenario
also calls for the vswitch=none flag. This patch resolves this issue
to stop the vdev devices from being added when the vnf is specified
as QemuPciPassthrough.

JIRA: VSPERF-460

Change-Id: Ie4522b046982ec5554c3deb6a507d86ca9798124
Signed-off-by: Christian Trautman <ctrautma@redhat.com>