Update patch set 1
Patch Set 1:

I still need to test this but its easier to test with the patch up in review so I can easily download it. My test systems are tied up atm so if anyone is available to verify they can so an SR-IOV test with a guest to make sure the code is correct it would be appreciated. Otherwise I'll test it in a few days.

Patch-set: 1