docs: Remove hyperlinks to docs in artifacts
It is possible to use sphinx cross referencing
feature in case, that only one level of toctree
is used. Thus documentation was reorganized
to utilize just one top level index with toctree.
After that it was possible to use rst native syntax
to create a cross links among vsperf documentation
chapters. As a result only 2 hyperlinks to the
yardstick documentation are version specific.

JIRA: VSPERF-442

Change-Id: I15c2f53922e88765506448de61af4a299f5d53c5
Signed-off-by: Martin Klozik <martinx.klozik@intel.com>
Reviewed-by: Al Morton <acmorton@att.com>
Reviewed-by: Christian Trautman <ctrautma@redhat.com>
Reviewed-by: Bill Michalowski <bmichalo@redhat.com>
Reviewed-by: Antonio Fischetti <antonio.fischetti@intel.com>
Reviewed-by: Sridhar Rao <sridhar.rao@spirent.com>
Reviewed-by: Martin Goldammer <martinx.goldammer@intel.com>
Reviewed-by: Trevor Cooper <trevor.cooper@intel.com>