Update patch set 3
Patch Set 3:

Hi Sridhar,

thank you for the review and a good question. According to my understanding, the search page is created automatically. Although I'm not sure if it is a generic sphinx feature or opnfvdocs specific implementation. So I've left original ref to search page untouched. It doesn't work in the docs generated for the review, but it should work once documentation is officially generated.

By the way, I would like to wait with this patch, until impact of move to read-the-docs will be clarified.

Best Regards,
Martin

Patch-set: 3
Reviewer: Gerrit User 1306 <1306@bbac25d2-bf60-4904-9ba8-a72fc000d6c5>
Label: Verified=0