Update patch set 2
Patch Set 2:

Regarding the IETF Draft and Martin Klozik's e-mail about the license, I don't think we want to add the IETF license to the .xml file.  I remember that we had the .txt version of the draft in our repo when we started, but were asked to delete the .txt files because of the IETF license, IIRC.  We could add the OPNFV license in an XML comment, but it cannot be the first line of the file like other license headers. It's getting close to when we could delete the XML files, they are very close to approval, but there is a delay in-progress, unfortunately.

Patch-set: 2