Update patch set 2
Patch Set 2:

Hi Al,

thanks for clarification of IETF draft xml files. From technical point of view, it is not a problem to put a license into the separate file in the same directory as Martin did it (i.e into ietf_draft/LICENSE). On the other hand, there is a legal part and if I got it correctly it has been already a problem to put IETF license into OPNFV repo. It would be probably a problem if we would put Creative Common license to IETF DRAFT, but final IETF document would have IETF license.

Thus I suggest to keep ietf draft documents without licenses for now. It is probably the best solution among (bad) options we have. Let's see if anybody will complain.

Patch-set: 2