Update patch set 1
Patch Set 1:

Hi Marco,

ad 1) ok, my fault, I should explicitly written, that in case of OvsVanilla you have to specify, that qemu should use virtionet to connect to OVS. I should pointed you to the doc to check section about OVS Vanilla testing. In a nutshell, you have to specify following parameters to the vsperf: --vswitch OvsVanilla --vnf QemuVirtioNet

ad 2) ok

ad 3&5) Could you update the documentation today? We have agreed with Trevor on Friday, that including you script in upcoming Danube release is safe, so we would like to merge it to the repo today.

ad 4) as it was not clear if you patch would made it to Danube, then I've pushed mine separately; So please remove it from your patch and do a rebase as part of 3)

Thanks, Martin

Patch-set: 1
Label: Code-Review=0