xena_verify: Add final verification option to Xena
Adds final verification option to Xena traffic generator
inside VSPerf. Works by running the Xena2544.exe again
after the initial run is complete with modified values
inside the json file. If the final verify fails it will
revert the settings and continue the in a pattern based
on if the smart search is enabled.

JIRA: VSPERF-506

Change-Id: Ic272628c489b88989f67549bd020af7354dc9b91
Signed-off-by: Christian Trautman <ctrautma@redhat.com>