Update patch set 1
Patch Set 1:

Great Work! I failed to write a comment in time, but the back2back tests use a Y-axis label "fps" and its should be "frames". It's also worth noting that a logarithmic Y axis will make the smaller values more visible, and this is important because the small values are the only useful information measured.

Patch-set: 1