build: Rename cloned repositories and remove cuse
VSPERF expects folders to be named a certain way, but this may not be
the case if another repository is specified via the package-list.mk
file. One example of this is the DPDK stable repo which gets cloned to
the "dpdk-stable" directory unless otherwise specified. Thus Makefiles
were updated to enforce expected folder names.

Unused configuration related to vHostCuse support was removed.

Change-Id: Ifa48898d51e96c72e7d740f43cb635d729b2a796
Signed-off-by: Cian Ferriter <cian.ferriter@intel.com>
Reviewed-by: Martin Klozik <martinx.klozik@intel.com>
Reviewed-by: Al Morton <acmorton@att.com>
Reviewed-by: Christian Trautman <ctrautma@redhat.com>
Reviewed-by: Sridhar Rao <sridhar.rao@spirent.com>
Reviewed-by: Trevor Cooper <trevor.cooper@intel.com>