Update patch set 2
Patch Set 2: Code-Review+1

I agree with Maryam's comments, *and* I'd like to see the patch merged to make this functionality possible. So, this is a two-step process. 
Thanks for tackling this, Martin! As Morgan said, it's much better to report the Success or Fail Status, as we see it 
(It's really Results or Fail, wording I'm ok with).

Patch-set: 2
Reviewer: Gerrit User 385 <385@bbac25d2-bf60-4904-9ba8-a72fc000d6c5>
Label: Code-Review=+1