Update patch set 4
Patch Set 4: Code-Review+2

Patch tested OK.

The only thing I'm seeing which is not related to this patch is running phy2phy with T-Rex is not generating enough packets at 64 bytes. It says 100% with 14033450 fps which is not correct for bi-directional traffic. Could be something in my config on the t-rex side so I'll investigate. If we can get multistream into this I think we will see more people move toward using this in the future. I know we will start using it more for public projects.

Patch-set: 4
Reviewer: Gerrit User 2851 <2851@bbac25d2-bf60-4904-9ba8-a72fc000d6c5>
Label: Code-Review=+2