Update patch set 1
Patch Set 1:

It has been decided to merge this patch as it is, i.e. with promiscuous mode turned OFF by default.

A separate ticket will be created for investigation of behaviour of other traffic generators. In the future (F release) we will synchronize promiscuous mode handling among all trafficgens and possibly prepare generic option to turn it of and on (e.g. TRAFFIC['promiscuous'] True/False.

Patch-set: 1