Update patch set 5
Patch Set 5:

Hi Martin,

So I think part of the confusion with the UDP ports is I'm only rotating the source port from both sides of the traffic. I do not touch the destination. What I'm doing on the second stream is using the destination port as the starting value. I think I'll change it to use the source port as well. I was in the mindset of how we use the destination IP as the starting point on the second stream for bi-direcitonal traffic, if that makes sense. I did change based on some of your suggestions. I'll get this going on the pods this morning so I can ask Sridhar to test and get his comments.

Patch-set: 5