Update patch set 7
Patch Set 7:

(1 comment)

I've put one final comment inline. Otherwise it looks good to me.

It would be great if you or Sridhar will be able to run a few test to verify its functionality, i.e.:

1) run phy2phy_cont & pvp_cont
2) run phy2phy_cont & pvp_cont with multistream set to 16000 to check that performance has been affected
3) run multistream for L2, L3 and L4, e.g. with 5 streams only and inspect traffic by wireshark (./vsperf --mode trafficgen ... )
4) (nice to have) L4 with mulstistream 5 and udp port 65534 - and inspect by wireshark, that ports from 65531 to 65535 are used
5) (nice to have) check phy2phy_cont with latency off & on
6) (nice to have) check phy2phy_cont with mulstistream 16000 with latency off & on

I would say that 1) and 2) are the minimum to be tested before merge.

Patch-set: 7
Label: Verified=0