Update patch set 1
Patch Set 1:

Thanks Christian, I was thinking about the wrong side of the learning frames, mea culpa.  But with the sleep at the end, we need to coordinate this 3 sec sleep time with the DUT's maximum address age time.  

Age Time - The maximum time that a DUT/SUT will keep a learned
      address in its forwarding table.

Patch-set: 1