Update patch set 11
Patch Set 11: Code-Review+2

It seems, that the reason of failing daily job is a server crash during reload of "opnvswitch" module. It seems, that it is related to kernel 3.10.xxx. In case that kernel 4.xx is used (together with newer OVS), then machine behaves more stable.

Thus I'm merging this patch, to get DAILY CI up and running after a while. We can always revert it or change tools versions if needed.

Patch-set: 11
Label: Code-Review=+2