Update patch set 2
Patch Set 2:

Hi Al,

thanks for your review. In case of test examples for traffic capture in VM or at HOST (in setup with 2 or more NICs) - the capture is not stored in permanent location. However it can be done in "real" tests. In case of VM it would require additional steps to mount shared disk, to store capture there and at the end of tc, this file have to be copied to results directory.

In case of second example - it is about stating a capture file name with full path into results directory (see Trex patch for notation).

In case of 3rd method - i.e. patch with traffic capture support in T-Rex - the capture files are stored automatically into directory with results and pcap (base)names are part of results (printed into vsperf output and stated in CSV file).

In all cases above are frames stored in PCAP file. However these are just examples, so tester can use a different tool (e.g. tshark, netcat, etc.) to capture and filter traffic inside VM or at DUT.

Examples provided in this patch are meant as an inspiration and demonstration of capabilities of step driven testcases.

--Martin

Patch-set: 2
Reviewer: Gerrit User 1306 <1306@bbac25d2-bf60-4904-9ba8-a72fc000d6c5>
Label: Verified=0