Update patch set 11
Patch Set 11:

> (2 comments)
 > 
 > a trivial comment and a question

Al,
The sample[2] and sample[3] contains the value and the time information, respectively. This is collected in line 194 in handle-function. Yes, it all starts from the collectd_bucky file, who returns the sample as - [Host, Name, Value and Time].
I have added a comment in the handle function to explain this.

Thanks!

Patch-set: 11