Update patch set 2
Patch Set 2:

Thanks for your comment, Maryam.  You're right to flag this requirement, in that we haven't had much success imposing requirements on HW T-gens, even to fix their bugs...   But, it seems highly likely that the Long Term tests at Linux Foundation pods will be using SW T-gens, with more sympathy to our needs.

In any case, I think the requirement should stand, as it reflects the feedback we received at IETF in Berlin, and incorporated in the RFC.

Patch-set: 2