Update patch set 7
Patch Set 7:

> (1 comment)
 > 
 > one small suggestion, otherwise looks good, thanks!

Hi Al,

The rate is the percentage to start at. I'll update the comment to reflect this. The idea is some tests such as OVS-Vanilla have such low bit rate results it may not be wise to start at 100% flow rates since it will never achieve this so to speed up testing you can specify a lower start rate. The other parts of this patch allow you to either let T-Rex API determine the maximum speed and this setting can control where to start that setting. Does that make sense? I'll fix the wording so this is more clear.

Patch-set: 7