Update patch set 8
Patch Set 8:

> (5 comments)
 > 
 > Looks good, thought I've not tested it myself. Just a few minor
 > comments.

Lastly on the forcing of speed, we would have to see if other traffic gens support doing this. The major reason for doing this with T-Rex is I've found if the speed requested on T-Rex can't be achieved due to card limitations or system limitations/configuration then the binary search acts very strangely. We can explore adding this to other t-gens later :)

Patch-set: 8
Label: Verified=0