Update patch set 1
Patch Set 1:

Is there any particular reason to add these two testcases? In general, vsperf supports many options and it is not possible to cover all permutations of these options by TC definitions. It is up to the tester to set appropriate options (e.g. sets of stream counts for multistream feature) to achieve testing goals.

In the past, the basic set of performance testcases was meant to:
1) implement a basic set of TCs for VSPERF initial use and execution of CI
2) provide an examples for various VSPERF features, so users can create their own custom testcases based on these examples

Patch-set: 1