Update patch set 1
Patch Set 1:

Hi Julien,

thanks for the patch. It is a step in the right direction. Could you please check and ensure if needed, that interfaces created by vsperf are also removed together with the bridges? For example Vanilla OVS creates TAP interfaces for QEMU's virtual NICs and these shall be removed at vsperf teardown too. There is also support for veth, etc, which should be checked to work analogically.

There is also an implementation of context manager of OFBridge class in src/ovs/ofctl.py. It was probably created to address the same issue as your patch, but it is not used(?). Could you double check, that this is a dead code and remove it as part of your patch?

I'm not sure how this patch was tested. Anyhow I would suggest to run several TCs at once (i.e. by one vsperf execution, e.g. phy2phy_cont pvp_cont pvvp_cont) for all supported vswitches (OVS with and without DPDK and VPP) and possibly for their combinations (e.g. pvp_cont and pvp_cont_vpp at the same vsperf execution). Please test also a few integration testcases, which removes bridge on their own, to ensure, that it doesn't cause any issue. Another improvement might be modification of explicit call of del_switch from integration tests, as it is going to be obsoleted by your patch.

Regards,
Martin

Patch-set: 1