Update patch set 1
Patch Set 1:

Hi Martin,

> 1) could you check if recent sysmetrics fixes pidstat cpu usage of a master process?

I will recompile with upstream version of sysstat and check it.

> 2) it should be noted, that cpu usage statistics calculated from "thread" statistics are not accurate;

Do you think that a log message with the level WARNING is enough for the user? 

> 3) values should be rounded to 2 decimal places to be consistent

I have also noticed it, I will rework it.

> 4) stored "results" should be always of the same type;

OK, I will always cast to str.

--
I have thinking a little more about this problem.

The main problem is that we cannot trust the version of pidstat in order to apply this workaround or not. For example, on Centos 7, version installed is 10.1.5 (10.1.5-12.el7), regression is introduced in 11.3.2 (vanilla), but cpu usage is wrong (regression backported).

As a fix is now available with the latest version of sysstat, the other workaround, if we want accurate statistics, is to build our own sysstat. What do you think ?

Regards,
Julien

Patch-set: 1
Label: Verified=0