Update patch set 6
Patch Set 6:

Updated the patch based on testing from Bob. We found the layer 4 mod code was not being executed because of an incorrect reference to a variable. We also modified layer 2 mods to only mod the last 4 sets of bits instead of the first 4. This prevents some invalid mac addresses with higher flow counts.

Patch-set: 6
Label: Verified=0